# openapi.yaml
openapi: 3.0.3
info:
  title: Kaldar Platform API Documentation
  description: |
    Complete Production REST API specification for the Kaldar Monorepo (Laundry & Dry Cleaning Platform).
    
    ### Full API Inventory (55 Endpoints across 14 Controllers):
    1. **Authentication & Verification**: Login, OTP verification & resending, Forgot Password, Reset Password.
    2. **Customer**: Registration, profile fetch, update profile, change password, delete customer account.
    3. **Dry Cleaner Store & Ops**: Registration, profile edit, CAC business verification & status check, service inventory management, payout bank account setup, working hours schedule, onboarding checklist status, analytics metrics, delete dry cleaner account.
    4. **Payout**: Bank account verification, earnings withdrawal to registered bank (Anchor Banking).
    5. **Discovery**: Geospatial discovery of nearby available dry cleaners.
    6. **Order Management**: Create order, accept order, reject order, order status progression, fetch order details, customer order history, dry cleaner order queue.
    7. **Review & Ratings**: Submit review, dry cleaner store reviews list, single order review fetch.
    8. **Logistics & Dispatch**: Delivery partner webhook callbacks (Chowdeck, Uber Direct, etc.).
    9. **Notification Center**: Paginated notifications list, toggle read status, delete single, bulk mark read/unread/delete, mark all read.
...
  version: 1.0.0
  contact:
    name: Kaldar Engineering Team
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/customer/register:
    post:
      tags: [Customer]
      summary: Register Customer Account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password, phoneNumber]
              properties:
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string }
                password: { type: string }
                phoneNumber: { type: string }
                address: { type: string }
      responses:
        '201': { description: Customer registered }
  /api/v1/customer/updated_customer_profile:
    put:
      tags: [Customer]
      summary: Update Customer Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                phoneNumber: { type: string }
                address: { type: string }
      responses:
        '200': { description: Profile updated }
  /api/v1/customer/{customerId}:
    get:
      tags: [Customer]
      summary: Get Customer Profile
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Profile retrieved }
    delete:
      tags: [Customer]
      summary: Delete Customer Account
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Customer account deleted successfully }
        '404': { description: Customer not found }
  /api/v1/customer/change_password:
    put:
      tags: [Customer]
      summary: Change Customer Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200': { description: Password updated }

  # =========================================================================
  # 3. DRY CLEANER & PAYOUT MODULE (12 Endpoints)
  # =========================================================================
...
  /api/v1/order/customer/{customerId}:
    get:
      tags: [Order]
      summary: Get Customer Order History
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Customer order list }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT